Dependency Injection again to DiagnosticoMecanicoDataLoader.cs and DiagnosticoMecanicoExcelLoader.cs
