+++ Infrastructure/Infrastructure.csproj
@@ -57,6 +57,9 @@
     <Reference Include="SkiaSharp.HarfBuzz, Version=2.88.0.0, Culture=neutral, PublicKeyToken=0738eb9f132ed756, processorArchitecture=MSIL">
       <HintPath>..\packages\SkiaSharp.HarfBuzz.2.88.7\lib\net462\SkiaSharp.HarfBuzz.dll</HintPath>
     </Reference>
+    <Reference Include="SpreadsheetLight">
+      <HintPath>..\packages\SpreadsheetLight.3.5.0\lib\netstandard2.0\SpreadsheetLight.dll</HintPath>
+    </Reference>
     <Reference Include="Svg.Custom, Version=1.0.0.18, Culture=neutral, PublicKeyToken=dafe96fe6c845a74, processorArchitecture=MSIL">
       <HintPath>..\packages\Svg.Custom.1.0.0.18\lib\netstandard2.0\Svg.Custom.dll</HintPath>
     </Reference>
@@ -99,12 +102,20 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConvertStatus.cs" />
+    <Compile Include="DiagnosticoMecanicoDataLoader.cs" />
+    <Compile Include="DiagnosticoMecanicoExcelLoader.cs" />
     <Compile Include="IPdfToWordServiceBase.cs" />
     <Compile Include="PdfToWordServiceImpl.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Domain\Domain.csproj">
+      <Project>{3eef8743-e3b8-4e59-b25c-e6ace66a073f}</Project>
+      <Name>Domain</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\HarfBuzzSharp.NativeAssets.macOS.7.3.0.1\build\net462\HarfBuzzSharp.NativeAssets.macOS.targets" Condition="Exists('..\packages\HarfBuzzSharp.NativeAssets.macOS.7.3.0.1\build\net462\HarfBuzzSharp.NativeAssets.macOS.targets')" />
